Change page 1 animation and background color
Added a big delay on the <hr> animation that transitions into the cards' animation. This means putting a long transition delay on all the cards but I made it short enough so that if the user goes to next page and back, it fully resets in time. 
Made all pages' background dark navy