Added scrolling function that doesn't work and changed threshold
Tried to change navbar color when scrolling to a new page but I can't get the distance scrolled unless I remove all overflow settings. I need overflow: scroll for the page snap so I'll just comment this out for later. 
Changed threshold so 99% needs to be seen for an element to be observed.